# services/auth/api/docs/swagger.yaml
definitions:
  github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientIdToken:
    properties:
      client_id:
        type: string
      given_time:
        type: string
      id:
        type: string
      info:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientPlatform:
    properties:
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CreateReq:
    properties:
      client_id:
        type: string
      client_platform_id:
        type: string
      client_type_id:
        type: string
      disable:
        type: boolean
      environment_id:
        type: string
      monthly_request_limit:
        type: integer
      name:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      rps_limit:
        type: integer
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CreateRes:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      client_id:
        type: string
      client_type_id:
        type: string
      created_at:
        type: string
      disable:
        type: boolean
      environment_id:
        type: string
      id:
        type: string
      monthly_request_limit:
        type: integer
      name:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      rps_limit:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.DeleteRes:
    properties:
      row_effected:
        type: integer
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GenerateApiTokenReq:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GenerateApiTokenRes:
    properties:
      token:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Token'
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetClientPlatformListResponse:
    properties:
      client_platforms:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientPlatform'
        type: array
      count:
        type: integer
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetListRes:
    properties:
      count:
        type: integer
      data:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetRes'
        type: array
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetRes:
    properties:
      app_id:
        type: string
      app_secret:
        type: string
      client_id:
        type: string
      client_platform:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientPlatform'
      client_type_id:
        type: string
      created_at:
        type: string
      disable:
        type: boolean
      environment_id:
        type: string
      id:
        type: string
      monthly_request_limit:
        type: integer
      name:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      rps_limit:
        type: integer
      status:
        type: string
      updated_at:
        type: string
      used_count:
        type: integer
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ListClientTokenResponse:
    properties:
      client_tokens:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ClientIdToken'
        type: array
      count:
        type: integer
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RefreshApiTokenReq:
    properties:
      refresh_token:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RefreshTokenRequest:
    properties:
      client_type_id:
        type: string
      env_id:
        type: string
      project_id:
        type: string
      refresh_token:
        type: string
      role_id:
        type: string
      tables:
        items:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.TableBody'
        type: array
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RefreshTokenResponse:
    properties:
      token:
        $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Token'
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.TableBody:
    properties:
      object_id:
        type: string
      table_slug:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.Token:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      expires_at:
        type: string
      refresh_in_seconds:
        type: integer
      refresh_token:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdateReq:
    properties:
      client_type_id:
        type: string
      id:
        type: string
      monthly_request_limit:
        type: integer
      name:
        type: string
      project_id:
        type: string
      role_id:
        type: string
      rps_limit:
        type: integer
      status:
        type: string
    type: object
  github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdateRes:
    properties:
      row_effected:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://u-code.io
paths:
...
  /refresh:
    put:
      consumes:
      - application/json
      description: Refresh Token
      operationId: refresh
      parameters:
      - description: RefreshTokenRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Refresh token data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RefreshTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Refresh Token
      tags:
      - Session
...
  /v2/api-key/{project-id}:
    get:
      consumes:
      - application/json
      description: Get List ApiKey
      operationId: get_list_api_key
      parameters:
      - description: project-id
        in: path
        name: project-id
        required: true
        type: string
      - description: environment-id
        in: query
        name: environment-id
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      - description: client_type_id
        in: query
        name: client_type_id
        type: string
      - description: role_id
        in: query
        name: role_id
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: ApiKey data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetListRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List ApiKey
      tags:
      - V2_ApiKey
    post:
      consumes:
      - application/json
      description: Create ApiKey
      operationId: create_api_keys
      parameters:
      - description: project-id
        in: path
        name: project-id
        required: true
        type: string
      - description: ApiKeyReqBody
        in: body
        name: api-key
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CreateReq'
      produces:
      - application/json
      responses:
        "201":
          description: ApiKey data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.CreateRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create ApiKey
      tags:
      - V2_ApiKey
  /v2/api-key/{project-id}/{id}:
    delete:
      consumes:
      - application/json
      description: Delete ApiKeys
      operationId: delete_api_keys
      parameters:
      - description: project-id
        in: path
        name: project-id
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: ApiKey data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.DeleteRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete ApiKeys
      tags:
      - V2_ApiKey
    get:
      consumes:
      - application/json
      description: Get ApiKey
      operationId: get_api_key
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: project-id
        in: path
        name: project-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: ApiKey data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ApiKey
      tags:
      - V2_ApiKey
    put:
      consumes:
      - application/json
      description: Update ApiKey
      operationId: update_api_keys
      parameters:
      - description: project-id
        in: path
        name: project-id
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: ApiKeyReqBody
        in: body
        name: api-key
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdateReq'
      produces:
      - application/json
      responses:
        "201":
          description: ApiKey data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.UpdateRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update ApiKey
      tags:
      - V2_ApiKey
  /v2/api-key/{project-id}/tokens:
    get:
      consumes:
      - application/json
      description: Get List Clint Tokens
      operationId: get_list_client_tokens
      parameters:
      - description: project-id
        in: path
        name: project-id
        required: true
        type: string
      - description: client-id
        in: query
        name: client-id
        type: string
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Tokens data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.ListClientTokenResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get List Clint Tokens
      tags:
      - V2_ApiKey_Token
  /v2/api-key/generate-token:
    post:
      consumes:
      - application/json
      description: Generate Api Key Token
      operationId: generate_api_key_token
      parameters:
      - description: ApiKeyReqBody
        in: body
        name: api-key
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GenerateApiTokenReq'
      produces:
      - application/json
      responses:
        "201":
          description: ApiKey data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GenerateApiTokenRes'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Generate Api Key Token
      tags:
      - V2_ApiKey
  /v2/api-key/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh Api Key Token
      operationId: refresh_api_key_token
      parameters:
      - description: ApiKeyReqBody
        in: body
        name: api-key
        required: true
        schema:
          $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RefreshApiTokenReq'
      produces:
      - application/json
      responses:
        "201":
          description: ApiKey data
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.RefreshApiTokenReq'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Refresh Api Key Token
      tags:
      - V2_ApiKey
...
  /v2/client-platform:
    get:
      consumes:
      - application/json
      description: Get ClientPlatform List
      operationId: get_client_platform_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetClientPlatformListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_genproto_auth_service.GetClientPlatformListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Ucode-io_ucode-opensource_services_auth_api_http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ClientPlatform List
      tags:
      - ClientPlatform
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"